Changes order to int float

+++ samples/openapi3/client/petstore/python/src/petstore_api/components/schema/array_of_array_of_number_only.py
@@ -15,7 +15,7 @@ Items2: typing_extensions.TypeAlias = schemas.NumberSchema
 
 class ItemsTuple(
     typing.Tuple[
-        typing.Union[float, int],
+        typing.Union[int, float],
         ...
     ]
 ):
