+++ samples/openapi3/client/petstore/python/src/petstore_api/components/schema/banana.py
@@ -22,11 +22,11 @@ Properties = typing_extensions.TypedDict(
 class BananaDict(schemas.immutabledict[str, schemas.OUTPUT_BASE_TYPES]):
     
     @property
-    def lengthCm(self) -> typing.Union[float, int]:
+    def lengthCm(self) -> typing.Union[int, float]:
         return self.__getitem__("lengthCm")
     
     @typing.overload
-    def __getitem__(self, name: typing_extensions.Literal["lengthCm"]) -> typing.Union[float, int]:
+    def __getitem__(self, name: typing_extensions.Literal["lengthCm"]) -> typing.Union[int, float]:
         ...
     
     @typing.overload
